services/media2text_service/requirements.txt
requests==2.32.4
fastapi==0.115.14
uvicorn==0.35.0
python-multipart
pydantic_settings==2.3.4
pydantic==2.8.2
openai-whisper==20250625
pytesseract==0.3.13
setuptools-rust==1.12.0
